% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Minimal hello world}
\usage{
hello_world(x = "world", return_val = TRUE)
}
\arguments{
\item{x}{Object to say hello to, defaults to "world"}

\item{return_val}{If true, return to new object, else print}
}
\value{
Optionally string, message output
}
\description{
Minimal hello world
}
